requirements/prod.txt
Django==4.2.16
aiogram==3.16.0
psycopg2-binary==2.9.9
python-decouple==3.8
